Start the separation of text renderer and drawing context
