+++ src/OpenLoco/CMakeLists.txt
@@ -151,6 +151,7 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/RenderTarget.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/SoftwareDrawingContext.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/SoftwareDrawingEngine.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/TextRenderer.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Gui.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Input.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Input/Keyboard.cpp"
@@ -482,6 +483,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/RenderTarget.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/SoftwareDrawingContext.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/SoftwareDrawingEngine.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Graphics/TextRenderer.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Gui.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Input.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Input/Shortcuts.h"
